Adjusted strafeLeft and runForwardsDistanceAndRaiseArm value
I think the main problem is the angle correction code, seems to not be working like it used to even though it was not touched. It would turn when not needed to. I will need to see if the distance sensors get the same numbers and the error it gets from it reads the correct number.
-Viassna